Fully split up apiRouytes into smaller files. Still need to test everything
