--- backend/apiRoutes.js
@@ -1,1670 +1,0 @@
-import dotenv from 'dotenv';
-import { createClient } from '@supabase/supabase-js';
-import express from 'express';
-import cors from 'cors';
-import bcrypt from 'bcrypt';
-import { v4 as uuidv4 } from 'uuid';
-
-dotenv.config();
-const router = express.Router();  
-
-// Supabase Client Initialization
-const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_KEY);
-
-router.use(cors());
-router.use(express.json());
-
-// lazy solution
-const TABLE_NAMES = [
-  "cort_gm",
-  "cort",
-  "function",
-  "function_test",
-  "gm",
-  "gm_function",
-  "reference",
-  "stimulation",
-  "tag",
-  "test",
-  "test_tag",
-  "electrode",
-  "localization",
-  "region_name"
-];
-
-// Fetch table names
-router.get("/tables", async (req, res) => {
-  res.json({ tables: TABLE_NAMES });
-});
-
-// Fetch data from a specific table
-router.get("/tables/:table", async (req, res) => {
-  const { table } = req.params;
-  
-  if (!TABLE_NAMES.includes(table)) {
-    return res.status(400).json({ error: "Invalid table name" });
-  }
-
-  try {
-    const { data, error } = await supabase.from(table).select("*").limit(100);
-    if (error) throw error;
-    res.json(data);
-  } catch (error) {
-    console.error("Server Error:", error);
-    res.status(500).json({ error: "Error fetching table data" });
-  }
-});
-
-
-// Fetch electrode label descriptions
-router.get("/electrode-label-descriptions", async (req, res) => {
-  try {
-    const { data, error } = await supabase
-      .from("electrode_label_description")
-      .select("*");
-    if (error) throw error;
-    res.status(200).json({ electrodeLabelDescriptions: data });
-  } catch (error) {
-    console.error("Error fetching electrode label descriptions:", error);
-    res.status(500).json({ error: error.message });
-  }
-});
-
-// Endpoint to save localization data
-router.post('/save-localization', async (req, res) => {
-  try {
-    console.log('Received localization save request', req.body);
-    
-    const { electrodes, fileId } = req.body;
-    
-    if (!electrodes) {
-      return res.status(400).json({ success: false, error: 'Missing electrodes data' });
-    }
-    
-    if (fileId === undefined || fileId === null) {
-      return res.status(400).json({ success: false, error: 'Missing file ID' });
-    }
-    
-    if (Object.keys(electrodes).length === 0) {
-      return res.status(400).json({ success: false, error: 'Electrodes data is empty' });
-    }
-    
-    console.log(`Processing save request for file ID: ${fileId} with ${Object.keys(electrodes).length} electrodes`);
-    
-    // Delete existing localization records for this file_id BEFORE saving new ones
-    console.log(`Checking for existing localizations with file_id: ${fileId}`);
-    const { data: existingData, error: checkError } = await supabase
-      .from('localization')
-      .select('id')
-      .eq('file_id', fileId);
-      
-    if (checkError) {
-      console.error('Error checking existing localizations:', checkError);
-    } else if (existingData && existingData.length > 0) {
-      console.log(`Found ${existingData.length} existing localizations with file_id ${fileId}, deleting...`);
-      const { error: deleteError } = await supabase
-        .from('localization')
-        .delete()
-        .eq('file_id', fileId);
-        
-      if (deleteError) {
-        console.error('Error deleting existing localizations:', deleteError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to update existing localization data: ${deleteError.message}`
-        });
-      }
-      console.log('Successfully deleted existing localizations');
-    } else {
-      console.log('No existing localizations found for this file_id');
-    }
-    
-    // Process the save operation with file ID
-    await saveLocalizationToDatabase(electrodes, fileId);
-    
-    res.status(200).json({ 
-      success: true,
-      message: 'Localization data saved successfully',
-      fileId: fileId
-    });
-  } catch (error) {
-    console.error('Error in save-localization endpoint:', error);
-    res.status(500).json({ 
-      success: false, 
-      error: error.message
-    });
-  }
-});
-
-/**
- * Handles file record management in the database.
- * Creates a new file record if it doesn't exist, or updates an existing one.
- * @param {string} fileId - The unique identifier for the file
- * @param {string} fileName - The name of the file
- * @param {string} creationDate - The creation date of the file
- * @param {string} modifiedDate - The last modified date of the file
- * @param {string} token - The authorization token for the user session
- * @returns {Promise<void>} - Resolves when the operation is complete, throws on error
- */
-async function handleFileRecord(fileId, fileName, creationDate, modifiedDate, token) {
-  // Check if file record already exists
-  const { data: existingFile } = await supabase
-    .from('files')
-    .select('*')
-    .eq('file_id', fileId)
-    .single();
-
-  if (existingFile) {
-    // Update existing file record
-    const { error: fileError } = await supabase
-      .from('files')
-      .update({
-        filename: fileName,
-        modified_date: modifiedDate
-      })
-      .eq('file_id', fileId);
-
-    if (fileError) throw fileError;
-  } else {
-    // Get user ID from session
-    if (!token) {
-      throw new Error('No authentication token provided');
-    }
-    
-    const { data: session } = await supabase
-      .from('sessions')
-      .select('user_id')
-      .eq('token', token)
-      .single();
-      
-    if (!session?.user_id) {
-      throw new Error('Invalid or expired session');
-    }
-
-    // Insert new file record
-    const { error: fileError } = await supabase
-      .from('files')
-      .insert({
-        file_id: fileId,
-        owner_user_id: session.user_id,
-        filename: fileName,
-        creation_date: creationDate,
-        modified_date: modifiedDate
-      });
-
-    if (fileError) throw fileError;
-  }
-}
-
-// Endpoint to save designation data
-router.post('/save-designation', async (req, res) => {
-  try {
-    console.log('Received designation save request', req.body);
-    
-    const { designationData, localizationData, fileId, fileName, creationDate, modifiedDate } = req.body;
-    
-    if (!designationData) {
-      return res.status(400).json({ success: false, error: 'Missing designation data' });
-    }
-    
-    if (!localizationData) {
-      return res.status(400).json({ success: false, error: 'Missing localization data' });
-    }
-    
-    if (fileId === undefined || fileId === null) {
-      return res.status(400).json({ success: false, error: 'Missing file ID' });
-    }
-
-    // Handle file record
-    try {
-      await handleFileRecord(fileId, fileName, creationDate, modifiedDate, req.headers.authorization);
-    } catch (error) {
-      console.error('Error saving file metadata:', error);
-      return res.status(500).json({ 
-        success: false, 
-        error: `Failed to save file metadata: ${error.message}`
-      });
-    }
-    
-    // Check for existing designation with this file_id
-    const { data: existingData, error: checkError } = await supabase
-      .from('designation')
-      .select('id')
-      .eq('file_id', fileId);
-      
-    if (checkError) {
-      console.error('Error checking existing designation:', checkError);
-    } else if (existingData && existingData.length > 0) {
-      console.log(`Found existing designation with file_id ${fileId}, updating...`);
-      const { error: updateError } = await supabase
-        .from('designation')
-        .update({
-          designation_data: designationData,
-          localization_data: localizationData
-        })
-        .eq('file_id', fileId);
-        
-      if (updateError) {
-        console.error('Error updating designation:', updateError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to update existing designation: ${updateError.message}`
-        });
-      }
-      console.log('Successfully updated designation');
-    } else {
-      // Insert new designation record
-      console.log('Creating new designation record...');
-      const { error: insertError } = await supabase
-        .from('designation')
-        .insert({
-          file_id: fileId,
-          designation_data: designationData,
-          localization_data: localizationData
-        });
-        
-      if (insertError) {
-        console.error('Error inserting designation:', insertError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to save designation: ${insertError.message}`
-        });
-      }
-      console.log('Successfully created new designation');
-    }
-    
-    res.status(200).json({ 
-      success: true,
-      message: 'Designation data saved successfully',
-      fileId: fileId
-    });
-  } catch (error) {
-    console.error('Error in save-designation endpoint:', error);
-    res.status(500).json({ 
-      success: false, 
-      error: error.message
-    });
-  }
-});
-
-// Endpoint to save stimulation data
-router.post('/save-stimulation', async (req, res) => {
-  try {
-    console.log('Received stimulation save request', req.body);
-    
-    const { electrodes, planOrder, isFunctionalMapping, fileId, fileName, creationDate, modifiedDate } = req.body;
-    
-    if (!electrodes) {
-      return res.status(400).json({ success: false, error: 'Missing electrodes data' });
-    }
-    
-    if (!planOrder) {
-      return res.status(400).json({ success: false, error: 'Missing plan order data' });
-    }
-    
-    if (fileId === undefined || fileId === null) {
-      return res.status(400).json({ success: false, error: 'Missing file ID' });
-    }
-    
-    if (isFunctionalMapping === undefined) {
-      return res.status(400).json({ success: false, error: 'Missing isFunctionalMapping flag' });
-    }
-    
-    console.log(`Processing stimulation save request for file ID: ${fileId}`);
-
-    // Handle file record
-    try {
-      await handleFileRecord(fileId, fileName, creationDate, modifiedDate, req.headers.authorization);
-    } catch (error) {
-      console.error('Error saving file metadata:', error);
-      return res.status(500).json({ 
-        success: false, 
-        error: `Failed to save file metadata: ${error.message}`
-      });
-    }
-    
-    // Check for existing stimulation with this file_id
-    const { data: existingData, error: checkError } = await supabase
-      .from('stimulation')
-      .select('id')
-      .eq('file_id', fileId);
-      
-    if (checkError) {
-      console.error('Error checking existing stimulation:', checkError);
-    } else if (existingData && existingData.length > 0) {
-      console.log(`Found existing stimulation with file_id ${fileId}, updating...`);
-      const { error: updateError } = await supabase
-        .from('stimulation')
-        .update({
-          stimulation_data: electrodes,
-          plan_order: planOrder,
-          is_mapping: isFunctionalMapping
-        })
-        .eq('file_id', fileId);
-        
-      if (updateError) {
-        console.error('Error updating stimulation:', updateError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to update existing stimulation: ${updateError.message}`
-        });
-      }
-      console.log('Successfully updated stimulation');
-    } else {
-      // Insert new stimulation record
-      console.log('Creating new stimulation record...');
-      const { error: insertError } = await supabase
-        .from('stimulation')
-        .insert({
-          file_id: fileId,
-          stimulation_data: electrodes,
-          plan_order: planOrder,
-          is_mapping: isFunctionalMapping
-        });
-        
-      if (insertError) {
-        console.error('Error inserting stimulation:', insertError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to save stimulation: ${insertError.message}`
-        });
-      }
-      console.log('Successfully created new stimulation');
-    }
-    
-    res.status(200).json({ 
-      success: true,
-      message: 'Stimulation data saved successfully',
-      fileId: fileId
-    });
-  } catch (error) {
-    console.error('Error in save-stimulation endpoint:', error);
-    res.status(500).json({ 
-      success: false, 
-      error: error.message
-    });
-  }
-});
-
-// Endpoint to save test selection data
-router.post('/save-test-selection', async (req, res) => {
-  try {
-    console.log('Received test selection save request', req.body);
-    
-    const { tests, contacts, fileId, fileName, creationDate, modifiedDate } = req.body;
-    
-    if (!tests) {
-      return res.status(400).json({ success: false, error: 'Missing tests data' });
-    }
-    
-    if (!contacts) {
-      return res.status(400).json({ success: false, error: 'Missing contacts data' });
-    }
-    
-    if (fileId === undefined || fileId === null) {
-      return res.status(400).json({ success: false, error: 'Missing file ID' });
-    }
-
-    // Handle file record
-    try {
-      await handleFileRecord(fileId, fileName, creationDate, modifiedDate, req.headers.authorization);
-    } catch (error) {
-      console.error('Error saving file metadata:', error);
-      return res.status(500).json({ 
-        success: false, 
-        error: `Failed to save file metadata: ${error.message}`
-      });
-    }
-    
-    // Check for existing test selection with this file_id
-    const { data: existingData, error: checkError } = await supabase
-      .from('test_selection')
-      .select('id')
-      .eq('file_id', fileId);
-      
-    if (checkError) {
-      console.error('Error checking existing test selection:', checkError);
-    } else if (existingData && existingData.length > 0) {
-      console.log(`Found existing test selection with file_id ${fileId}, updating...`);
-      const { error: updateError } = await supabase
-        .from('test_selection')
-        .update({
-          tests: tests,
-          contacts: contacts
-        })
-        .eq('file_id', fileId);
-        
-      if (updateError) {
-        console.error('Error updating test selection:', updateError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to update existing test selection: ${updateError.message}`
-        });
-      }
-      console.log('Successfully updated test selection');
-    } else {
-      // Insert new test selection record
-      console.log('Creating new test selection record...');
-      const { error: insertError } = await supabase
-        .from('test_selection')
-        .insert({
-          file_id: fileId,
-          tests: tests,
-          contacts: contacts
-        });
-        
-      if (insertError) {
-        console.error('Error inserting test selection:', insertError);
-        return res.status(500).json({ 
-          success: false, 
-          error: `Failed to save test selection: ${insertError.message}`
-        });
-      }
-      console.log('Successfully created new test selection');
-    }
-    
-    res.status(200).json({ 
-      success: true,
-      message: 'Test selection data saved successfully',
-      fileId: fileId
-    });
-  } catch (error) {
-    console.error('Error in save-test-selection endpoint:', error);
-    res.status(500).json({ 
-      success: false, 
-      error: error.message
-    });
-  }
-});
-
-router.post("/search", async (req, res) => {
-  try {
-    const { query, hemisphere, lobe } = req.body;
-
-    // Initialize result structure
-    const results = {
-      cort: [],
-      gm: [],
-      functions: [],
-      tests: []
-    };
-
-    // Search cort table with filters
-    let cortQuery = supabase
-      .from('cort')
-      .select(`
-        *,
-        cort_gm(
-          gm(
-            *,
-            cort_gm(
-              cort(*)
-            ),
-            gm_function(
-              function(*)
-            )
-          )
-        )
-      `);
-
-    // filtering
-    if (query) {
-      cortQuery = cortQuery.or(`name.ilike.%${query}%,acronym.ilike.%${query}%,lobe.ilike.%${query}%`);
-    }
-    if (hemisphere?.length) {
-      cortQuery = cortQuery.in('hemisphere', hemisphere.map(h => h === 'left' ? 'l' : 'r'));
-    }
-    if (lobe?.length) {
-      cortQuery = cortQuery.or(lobe.map(l => `lobe.ilike.%${l}%`).join(','));
-    }
-
-    const { data: cortData, error: cortError } = await cortQuery;
-    if (cortError) throw cortError;
-    results.cort = cortData || [];
-
-    // Process cort data and collect GM ids
-    results.cort.forEach(cort => {
-      cort.cort_gm?.forEach(cg => {
-        if (cg.gm && !results.gm.some(g => g.id === cg.gm.id)) {
-          results.gm.push({
-            ...cg.gm,
-            cort_gm: [{ cort: cort }]
-          });
-        }
-      });
-    });
-
-    // Combined search for GM, functions, and tests when query exists
-    if (query) {
-      // Search GM, functions, and tests in parallel
-      const [gmData, funcData, testData] = await Promise.all([
-        supabase
-          .from('gm')
-          .select(`
-            *,
-            cort_gm(
-              cort(*)
-            ),
-            gm_function(
-              function(*)
-            )
-          `)
-          .or(`name.ilike.%${query}%,acronym.ilike.%${query}%`),
-
-        supabase
-          .from('function')
-          .select(`
-            *,
-            gm_function(
-              gm(*)
-            ),
-            function_test(
-              test(*)
-            )
-          `)
-          .or(`name.ilike.%${query}%,description.ilike.%${query}%`),
-
-        supabase
-          .from('test')
-          .select(`
-            *,
-            function_test(
-              function(*)
-            )
-          `)
-          .or(`name.ilike.%${query}%,description.ilike.%${query}%`)
-      ]);
-
-      // Process test results
-      const relatedFuncIds = new Set();
-      if (!testData.error) {
-        testData.data?.forEach(test => {
-          if (!results.tests.some(t => t.id === test.id)) {
-            results.tests.push(test);
-            // Collect Function ids from function_test
-            test.function_test?.forEach(ft => {
-              if (ft.function?.id) relatedFuncIds.add(ft.function.id);
-            })
-          }
-        });
-      }
-
-      // Process function results and collect related GM ids
-      const relatedGmIds = new Set();
-      if (!funcData.error) {
-        funcData.data?.forEach(func => {
-          if (!results.functions.some(f => f.id === func.id)) {
-            results.functions.push(func);
-            // Collect GM ids from function relationships
-            func.gm_function?.forEach(gf => {
-              if (gf.gm?.id) relatedGmIds.add(gf.gm.id);
-            });
-          }
-        });
-      }
-
-      // Process GM results
-      if (!gmData.error) {
-        gmData.data?.forEach(gm => {
-          if (!results.gm.some(g => g.id === gm.id)) {
-            results.gm.push(gm);
-          }
-        });
-      }
-
-      // Fetch any missing Functions that are related to found tests
-      if (relatedFuncIds.size > 0) {
-        const funcIdsToFetch = Array.from(relatedFuncIds).filter(id =>
-          !results.gm.some(g => g.id === id)
-        );
-
-        if (funcIdsToFetch.length > 0) {
-          const { data: missingFuncs, error: missingFuncError } = await supabase
-            .from('function')
-            .select(`
-              *,
-              function_test(
-                test(*)
-              ),
-              gm_function(
-                function(*)
-              )
-            `)
-            .in('id', funcIdsToFetch);
-
-          if (!missingFuncError) {
-            missingFuncs?.forEach(func => {
-              if (!results.functions.some(f => f.id === func.id)) {
-                results.functions.push(func);
-              }
-            });
-          }
-        }
-      }
-
-      // Fetch any missing GMs that are related to found functions
-      if (relatedGmIds.size > 0) {
-        const gmIdsToFetch = Array.from(relatedGmIds).filter(id =>
-          !results.gm.some(g => g.id === id)
-        );
-
-        if (gmIdsToFetch.length > 0) {
-          const { data: missingGms, error: missingGmError } = await supabase
-            .from('gm')
-            .select(`
-              *,
-              cort_gm(
-                cort(*)
-              ),
-              gm_function(
-                function(*)
-              )
-            `)
-            .in('id', gmIdsToFetch);
-
-          if (!missingGmError) {
-            missingGms?.forEach(gm => {
-              if (!results.gm.some(g => g.id === gm.id)) {
-                results.gm.push(gm);
-              }
-            });
-          }
-        }
-      }
-    }
-
-    // Get all unique GM ids now (from cort, direct search, and function relationships)
-    const allGmIds = [...new Set(results.gm.map(g => g.id))];
-
-
-
-    // Fetch all relationships in parallel
-    // cort from GM and function from GM
-    const [gmFunctions, cortGM] = await Promise.all([
-      allGmIds.length > 0 ? supabase
-        .from('gm_function')
-        .select(`
-          *,
-          function(
-            *,
-            gm_function(
-              gm(*)
-            ),
-            function_test(
-              test(*)
-            )
-          )
-        `)
-        .in('gm_id', allGmIds) : { data: null, error: null },
-
-      allGmIds.length > 0 ? supabase
-        .from('cort_gm')
-        .select(`
-          *,
-          cort(
-            *,
-            cort_gm(
-              gm(*)
-            )
-          )
-        `)
-        .in('gm_id', allGmIds) : { data: null, error: null },
-    ]);
-
-    // Process gm_function relationships
-    if (!gmFunctions.error && gmFunctions.data) {
-      gmFunctions.data.forEach(gf => {
-        if (gf.function && !results.functions.some(f => f.id === gf.function.id)) {
-          results.functions.push(gf.function);
-        }
-      });
-    }
-
-    // Process cort_gm relationships
-    if (!cortGM.error && cortGM.data) {
-      cortGM.data.forEach(cg => {
-        if (cg.cort && !results.cort.some(c => c.id === cg.cort.id)) {
-          results.cort.push(cg.cort);
-        }
-      });
-    }
-
-    // Get all unique function ids
-    const allFunctionIds = [...new Set(results.functions.map(f => f.id))];
-
-    // test from function
-    const [functionTests] = await Promise.all([
-      allFunctionIds.length > 0 ? supabase
-        .from('function_test')
-        .select(`
-          *,
-          test(
-            *,
-            function_test(
-              function(*)
-            )
-          )
-        `)
-        .in('function_id', allFunctionIds) : { data: null, error: null },
-    ])
-
-    // Process function_test relationships
-    if (!functionTests.error && functionTests.data) {
-      functionTests.data.forEach(ft => {
-        if (ft.test && !results.tests.some(t => t.id === ft.test.id)) {
-          results.tests.push(ft.test);
-        }
-      });
-    }
-
-    // Filter relationships to only include items that exist in results
-    results.gm.forEach(gm => {
-      if (gm.cort_gm) {
-        gm.cort_gm = gm.cort_gm.filter(cg =>
-          results.cort.some(c => c.id === cg.cort?.id)
-        );
-      }
-      if (gm.gm_function) {
-        gm.gm_function = gm.gm_function.filter(gf =>
-          results.functions.some(f => f.id === gf.function?.id)
-        );
-      }
-    });
-
-    results.functions.forEach(func => {
-      if (func.gm_function) {
-        func.gm_function = func.gm_function.filter(gf =>
-          results.gm.some(g => g.id === gf.gm?.id)
-        );
-      }
-      if (func.function_test) {
-        func.function_test = func.function_test.filter(ft =>
-          results.tests.some(t => t.id === ft.test?.id)
-        );
-      }
-    });
-
-    results.tests.forEach(test => {
-      if (test.function_test) {
-        test.function_test = test.function_test.filter(ft =>
-          results.functions.some(f => f.id === ft.function?.id)
-        );
-      }
-    });
-
-    res.json(results);
-  } catch (error) {
-    console.error("Search Error:", error);
-    res.status(500).json({ error: "Error performing search" });
-  }
-});
-
-router.post('/suggest', async (req, res) => {
-    try {
-        const { query } = req.body;
-
-        if (!query) {
-            return res.status(400).json({ error: 'Query parameters are required' });
-        }
-
-        const { data: cortData, error: cortError } = await supabase
-            .from('cort')
-            .select('lobe, name, acronym')
-            .or(`lobe.ilike.%${query}%,name.ilike.%${query}%,acronym.ilike.%${query}%`)
-            .limit(15);
-
-        if (cortError) throw cortError;
-
-        const { data: gmData, error: gmError } = await supabase
-            .from('gm')
-            .select('name, acronym')
-            .or(`name.ilike.%${query}%,acronym.ilike.%${query}%`)
-            .limit(10);
-
-        if (gmError) throw gmError;
-
-        const suggestions = new Set();
-        cortData.forEach(item => {
-            if (item.lobe?.includes(query)) suggestions.add(item.lobe);
-            if (item.name?.includes(query)) suggestions.add(item.name);
-            if (item.acronym?.includes(query)) suggestions.add(item.acronym);
-        });
-        gmData.forEach(item => {
-            if (item.name?.includes(query)) suggestions.add(item.name);
-            if (item.acronym?.includes(query)) suggestions.add(item.acronym);
-        });
-
-        res.json({ suggestions: [...suggestions].slice(0, 25) });
-    } catch (err) {
-        console.error("Suggestion error:", err);
-        res.status(500).json({ error: 'Failed to fetch suggestions' });
-    }
-});
-
-router.get("/lobe-options", async (req, res) => {
-  try {
-    const { data, error } = await supabase
-      .from("cort")
-      .select("lobe")
-      .not("lobe", "is", null)
-      .not("lobe", "eq", "")
-      .order("lobe", { ascending: true });
-
-    if (error) throw error;
-
-    const uniqueLobes = [...new Set(data.map(item => item.lobe))].filter(Boolean);
-
-    res.json({ lobes: uniqueLobes });
-  } catch (error) {
-    console.error("Server Error:", error);
-    res.status(500).json({ error: "Error fetching lobe options" });
-  }
-});
-
-/**
- * Generates an acronym from a description.
- * @param {string} description - The description to generate an acronym from.
- * @returns {string} - The generated acronym.
- */
-function generateAcronym(description) {
-  return description
-    .split('')
-    .filter(char => char >= 'A' && char <= 'Z')
-    .join('');
-}
-
-/**
- * Inserts new regions into the region_name table and returns a mapping of region names to their IDs.
- * @param {string[]} regions - Array of region names.
- * @returns {Object} - A mapping of region names to their IDs.
- */
-async function insertRegionsAndGetIds(regions) {
-  const regionIdMap = {};
-
-  try {
-    // Fetch existing regions
-    const { data: existingRegions, error: fetchError } = await supabase
-      .from('region_name')
-      .select('id, name');
-
-    if (fetchError) {
-      console.error('Error fetching existing regions:', fetchError);
-      console.error('Error details:', JSON.stringify(fetchError));
-      throw fetchError;
-    }
-
-    // Create a map of lowercase names to their database versions and IDs
-    const existingNamesMap = new Map();
-    existingRegions.forEach(region => {
-      const lowerName = region.name.toLowerCase();
-      existingNamesMap.set(lowerName, {
-        id: region.id,
-        name: region.name
-      });
-    });
-
-    // Filter out regions that already exist (case-insensitive)
-    const newRegionsSet = new Set();
-    const newRegions = regions.filter(region => {
-      if (!region) return false;
-      const lowerRegion = region.toLowerCase();
-      
-      // If this region (case-insensitive) already exists
-      if (existingNamesMap.has(lowerRegion)) {
-        // Use the existing version from the database
-        const existing = existingNamesMap.get(lowerRegion);
-        regionIdMap[existing.name] = existing.id; // Only map the database version
-        return false;
-      }
-      
-      // If we haven't seen this region (case-insensitive) in this batch
-      if (!newRegionsSet.has(lowerRegion)) {
-        newRegionsSet.add(lowerRegion);
-        return true;
-      }
-      
-      return false;
-    });
-
-    if (newRegions.length > 0) {
-      console.log(`Inserting ${newRegions.length} new regions:`, newRegions);
-      
-      // Insert new regions
-      const { data: insertedRegions, error: insertError } = await supabase
-        .from('region_name')
-        .insert(newRegions.map(name => ({ name })))
-        .select();
-
-      if (insertError) {
-        console.error('Error inserting new regions:', insertError);
-        console.error('Error details:', JSON.stringify(insertError));
-        console.error('New regions data:', newRegions);
-        throw insertError;
-      }
-
-      console.log(`Successfully inserted ${insertedRegions.length} new regions`);
-      
-      // Add new regions to the map
-      insertedRegions.forEach(region => {
-        regionIdMap[region.name] = region.id;
-      });
-    }
-
-    return regionIdMap;
-  } catch (error) {
-    console.error('Error in insertRegionsAndGetIds:', error);
-    console.error('Error stack:', error.stack);
-    throw error;
-  }
-}
-
-/**
- * Saves localization data to the database in the format of the three tables.
- * @param {Object} data - The nested dictionary data from parseLocalization.
- * @param {number|string} fileId - The unique file identifier
- */
-async function saveLocalizationToDatabase(data, fileId) {
-  try {
-    console.log(`Starting save process for file ID: ${fileId}`);
-    
-    // Ensure fileId is an integer
-    const numericFileId = typeof fileId === 'string' ? parseInt(fileId) : fileId;
-    
-    if (isNaN(numericFileId)) {
-      throw new Error(`Invalid file ID: ${fileId} (must be a valid number)`);
-    }
-    
-    console.log(`Saving ${Object.keys(data).length} electrodes for file ID: ${numericFileId}`);
-    console.log('Sample data structure:', JSON.stringify(Object.entries(data)[0], null, 2));
-    
-    // Insert into electrode table
-    const electrodeData = Object.keys(data).map(label => ({
-      acronym: generateAcronym(data[label].description),  // Generate acronym from description
-      description: data[label].description, // Use description as electrode_desc
-      contact_number: Object.keys(data[label]).filter(key => key !== 'description' && key !== 'type').length, // Subtract description and type keys
-      label: label,
-      type: data[label].type || 'DIXI' // Include the electrode type, default to DIXI if not specified
-    }));
-
-    console.log(`Inserting ${electrodeData.length} electrodes...`);
-    
-    // Added detailed error checking for electrode insert
-    const { data: insertedElectrodes, error: electrodeError } = await supabase
-      .from('electrode')
-      .insert(electrodeData)
-      .select();
-
-    if (electrodeError) {
-      console.error('Error inserting electrodes:', electrodeError);
-      console.error('Error details:', JSON.stringify(electrodeError));
-      console.error('Sample electrode data:', JSON.stringify(electrodeData[0]));
-      throw electrodeError;
-    }
-
-    console.log(`Successfully inserted ${insertedElectrodes.length} electrodes`);
-
-    // Create a mapping of electrode labels to their database IDs
-    const electrodeIdMap = {};
-    insertedElectrodes.forEach(electrode => {
-      electrodeIdMap[electrode.label] = electrode.id;
-    });
-
-    console.log('Electrode ID mapping:', electrodeIdMap);
-
-    // Collect all unique regions from the data
-    const uniqueRegions = new Set();
-    Object.values(data).forEach(contacts => {
-      Object.values(contacts).forEach(contactData => {
-        if (typeof contactData === 'object' && contactData.associatedLocation) {
-          if (contactData.associatedLocation === 'GM/GM') {
-            const [desc1, desc2] = contactData.contactDescription.split('+');
-            uniqueRegions.add(desc1);
-            uniqueRegions.add(desc2);
-          }
-          else if (contactData.associatedLocation === 'GM' || contactData.associatedLocation === 'GM/WM') {
-            uniqueRegions.add(contactData.contactDescription);
-          }
-        }
-      });
-    });
-
-    console.log(`Found ${uniqueRegions.size} unique regions`);
-
-    // Insert new regions and get a mapping of region names to IDs
-    const regionIdMap = await insertRegionsAndGetIds(Array.from(uniqueRegions));
-    
-    console.log('Region ID mapping:', regionIdMap);
-
-    // Get the highest existing ID in the localization table
-    const { data: maxIdData, error: maxIdError } = await supabase
-      .from('localization')
-      .select('id')
-      .order('id', { ascending: false })
-      .limit(1);
-
-    if (maxIdError) {
-      console.error('Error getting max ID:', maxIdError);
-      console.error('Error details:', JSON.stringify(maxIdError));
-      throw maxIdError;
-    }
-
-    let nextId = maxIdData.length > 0 ? maxIdData[0].id + 1 : 1;
-    console.log(`Starting localization IDs from ${nextId}`);
-
-    // Insert into localization table with file_id
-    const localizationData = [];
-    console.log('DEBUG: Starting to build localization records from data structure:');
-    console.log('DEBUG: Electrode count:', Object.keys(data).length);
-    console.log('DEBUG: Sample electrode data:', JSON.stringify(Object.entries(data)[0], null, 2));
-    
-    let skippedDescriptionCount = 0;
-    let skippedNonObjectCount = 0;
-    let skippedNoLocationCount = 0;
-    let processedContactCount = 0;
-    
-    Object.entries(data).forEach(([label, contacts]) => {
-      console.log(`DEBUG: Processing electrode "${label}" with ${Object.keys(contacts).length} keys`);
-      console.log(`DEBUG: Contact keys:`, Object.keys(contacts));
-      
-      Object.entries(contacts).forEach(([contactNumber, contactData]) => {
-        console.log(`DEBUG: Processing contact "${contactNumber}" for electrode "${label}"`);
-        
-        if (contactNumber === 'description') {
-          console.log(`DEBUG: Skipping "description" key`);
-          skippedDescriptionCount++;
-          return; // Skip the description key
-        }
-        
-        if (typeof contactData !== 'object') {
-          console.log(`DEBUG: Skipping non-object contactData:`, contactData);
-          skippedNonObjectCount++;
-          return; // Skip invalid entries
-        }
-        
-        if (!contactData.associatedLocation || contactData.associatedLocation === '') {
-          console.log(`DEBUG: Setting default 'WM' location for contact without associatedLocation: ${label}/${contactNumber}`);
-          console.log(`DEBUG: Original contactData:`, contactData);
-          // Instead of skipping, assign default 'WM' as associatedLocation
-          contactData.associatedLocation = 'WM';
-          
-          // If there's no contactDescription either, set a default value
-          if (!contactData.contactDescription || contactData.contactDescription === '') {
-            contactData.contactDescription = 'Unknown';
-          }
-        }
-        
-        processedContactCount++;
-        console.log(`DEBUG: Valid contact found - ${label}/${contactNumber}: ${contactData.associatedLocation}`);
-        
-        const { contactDescription, associatedLocation } = contactData;
-        
-        if (!electrodeIdMap[label]) {
-          console.error(`DEBUG: ERROR - No electrode ID found for label "${label}"`);
-          console.log('DEBUG: Available electrode IDs:', electrodeIdMap);
-          return; // Skip if no electrode ID found
-        }
-        
-        // Handle GM/GM case: Split into two entries
-        if (associatedLocation === 'GM/GM') {
-          const [desc1, desc2] = contactDescription.split('+');
-          
-          if (!regionIdMap[desc1]) {
-            console.error(`DEBUG: ERROR - No region ID found for "${desc1}"`);
-            console.log('DEBUG: Available region IDs:', regionIdMap);
-            return;
-          }
-          
-          if (!regionIdMap[desc2]) {
-            console.error(`DEBUG: ERROR - No region ID found for "${desc2}"`);
-            console.log('DEBUG: Available region IDs:', regionIdMap);
-            return;
-          }
-          
-          console.log(`DEBUG: Adding two GM/GM records for ${label}/${contactNumber} (${desc1}, ${desc2})`);
-          
-          localizationData.push(
-            {
-              id: nextId++,
-              electrode_id: electrodeIdMap[label],
-              contact: contactNumber,
-              tissue_type: 'GM',
-              region_id: regionIdMap[desc1], // Use region ID from the map
-              file_id: numericFileId // Add file ID to localization record
-            },
-            {
-              id: nextId++,
-              electrode_id: electrodeIdMap[label],
-              contact: contactNumber,
-              tissue_type: 'GM',
-              region_id: regionIdMap[desc2], // Use region ID from the map
-              file_id: numericFileId // Add file ID to localization record
-            }
-          );
-        } else {
-          if (!regionIdMap[contactDescription]) {
-            console.error(`DEBUG: ERROR - No region ID found for "${contactDescription}"`);
-            console.log('DEBUG: Available region IDs:', regionIdMap);
-            return;
-          }
-          
-          console.log(`DEBUG: Adding record for ${label}/${contactNumber} (${associatedLocation}/${contactDescription})`);
-          
-          localizationData.push({
-            id: nextId++,
-            electrode_id: electrodeIdMap[label],
-            contact: contactNumber,
-            tissue_type: associatedLocation,
-            region_id: regionIdMap[contactDescription], // Use region ID from the map
-            file_id: numericFileId // Add file ID to localization record
-          });
-        }
-      });
-    });
-    
-    console.log('DEBUG: Localization data build summary:');
-    console.log(`DEBUG: - Skipped description keys: ${skippedDescriptionCount}`);
-    console.log(`DEBUG: - Skipped non-object data: ${skippedNonObjectCount}`);
-    console.log(`DEBUG: - Skipped no location data: ${skippedNoLocationCount}`);
-    console.log(`DEBUG: - Processed valid contacts: ${processedContactCount}`);
-    console.log(`DEBUG: - Final record count: ${localizationData.length}`);
-
-    if (localizationData.length === 0) {
-      console.warn('No localization data to insert');
-      
-      // Remove the test insertion that's likely causing errors
-      // Just log a warning instead
-      console.log('DEBUG: No localization data to save for this file');
-      return;
-    }
-
-    console.log(`Inserting ${localizationData.length} localization records with file_id: ${numericFileId}`);
-    console.log('Sample localization record:', JSON.stringify(localizationData[0], null, 2));
-    
-    // Try inserting in smaller batches to identify problem records
-    try {
-      // Remove the single test record insertion - this could be causing the error
-      
-      // Proceed with full batch insertion directly
-      console.log('Proceeding with batch insertion');
-      
-      const { data: insertedRecords, error: localizationError } = await supabase
-        .from('localization')
-        .insert(localizationData)
-        .select();
-
-      if (localizationError) {
-        console.error('Error inserting localization records:', localizationError);
-        console.error('Error details:', JSON.stringify(localizationError));
-        console.error('Data sample:', JSON.stringify(localizationData.slice(0, 3)));
-        
-        // Check the table structure to help diagnose
-        const { data: tableInfo, error: infoError } = await supabase
-          .from('localization')
-          .select('*')
-          .limit(1);
-          
-        if (infoError) {
-          console.error('Error getting table structure:', infoError);
-        } else {
-          console.log('Existing table structure sample:', tableInfo);
-        }
-        
-        throw localizationError;
-      }
-
-      console.log(`Successfully inserted ${localizationData.length} localization records`);
-      
-    } catch (batchError) {
-      console.error('Batch insertion failed:', batchError);
-      
-      // Try inserting records one by one to identify problematic records
-      console.log('Attempting record-by-record insertion to identify problem...');
-      
-      for (let i = 0; i < localizationData.length; i++) {
-        try {
-          const { error: singleError } = await supabase
-            .from('localization')
-            .insert([localizationData[i]])
-            .select();
-            
-          if (singleError) {
-            console.error(`Error at record ${i}:`, singleError);
-            console.error('Problematic record:', JSON.stringify(localizationData[i], null, 2));
-          }
-        } catch (singleRecordError) {
-          console.error(`Exception at record ${i}:`, singleRecordError);
-        }
-      }
-      
-      // Re-throw the original error
-      throw batchError;
-    }
-    
-    // Verify records were saved by counting
-    const { count, error: countError } = await supabase
-      .from('localization')
-      .select('*', { count: 'exact', head: true })
-      .eq('file_id', numericFileId);
-      
-    if (countError) {
-      console.warn('Error verifying saved records:', countError);
-    } else {
-      console.log(`Verification: Found ${count} records with file_id ${numericFileId}`);
-      if (count === 0) {
-        console.error('CRITICAL: Records appear to have been saved but count is 0');
-      }
-    }
-
-    return {
-      electrodeCount: insertedElectrodes.length,
-      localizationCount: localizationData.length,
-      fileId: numericFileId
-    };
-  } catch (error) {
-    console.error("Error saving data to the database:", error);
-    console.error("Error stack:", error.stack);
-    
-    // Check database connection
-    try {
-      const { data, error: pingError } = await supabase
-        .from('localization')
-        .select('count(*)', { count: 'exact', head: true });
-        
-      if (pingError) {
-        console.error('Database connection issue:', pingError);
-      } else {
-        console.log('Database connection is working');
-      }
-    } catch (pingError) {
-      console.error('Failed to ping database:', pingError);
-    }
-    
-    throw error; // Re-throw to propagate to the API endpoint handler
-  }
-}
-
-// User Profile Endpoints
-router.get("/user/profile", async (req, res) => {
-  const token = req.headers.authorization?.split(' ')[1];
-  if (!token) {
-    return res.status(401).json({ error: "No authentication token provided" });
-  }
-
-  try {
-    const { data: session } = await supabase
-      .from('sessions')
-      .select('user_id')
-      .eq('token', token)
-      .single();
-
-    if (!session?.user_id) {
-      return res.status(401).json({ error: "Invalid or expired session" });
-    }
-
-    const { data: user, error } = await supabase
-      .from('users')
-      .select('name')
-      .eq('id', session.user_id)
-      .single();
-
-    if (error) throw error;
-    res.json(user);
-  } catch (error) {
-    console.error('Error fetching user profile:', error);
-    res.status(500).json({ error: "Error fetching user profile" });
-  }
-});
-
-// File Operations Endpoints
-router.get("/files", async (req, res) => {
-  const token = req.headers.authorization?.split(' ')[1];
-  if (!token) {
-    return res.status(401).json({ error: "No authentication token provided" });
-  }
-
-  try {
-    const { data: session } = await supabase
-      .from('sessions')
-      .select('user_id')
-      .eq('token', token)
-      .single();
-
-    if (!session?.user_id) {
-      return res.status(401).json({ error: "Invalid or expired session" });
-    }
-
-    const { data: files, error } = await supabase
-      .from('files')
-      .select('*')
-      .eq('owner_user_id', session.user_id)
-      .order('modified_date', { ascending: false });
-
-    if (error) throw error;
-    res.json(files || []);
-  } catch (error) {
-    console.error('Error fetching user files:', error);
-    res.status(500).json({ error: "Error fetching user files" });
-  }
-});
-
-router.post("/files/metadata", async (req, res) => {
-  const token = req.headers.authorization?.split(' ')[1];
-  if (!token) {
-    return res.status(401).json({ error: "No authentication token provided" });
-  }
-
-  const { fileId, fileName, creationDate, modifiedDate } = req.body;
-
-  try {
-    const { data: session } = await supabase
-      .from('sessions')
-      .select('user_id')
-      .eq('token', token)
-      .single();
-
-    if (!session?.user_id) {
-      return res.status(401).json({ error: "Invalid or expired session" });
-    }
-
-    const { data: existingFile } = await supabase
-      .from('files')
-      .select('*')
-      .eq('file_id', fileId)
-      .single();
-
-    if (existingFile) {
-      const { error } = await supabase
-        .from('files')
-        .update({
-          filename: fileName,
-          modified_date: modifiedDate
-        })
-        .eq('file_id', fileId);
-
-      if (error) throw error;
-    } else {
-      const { error } = await supabase
-        .from('files')
-        .insert({
-          file_id: fileId,
-          owner_user_id: session.user_id,
-          filename: fileName,
-          creation_date: creationDate,
-          modified_date: modifiedDate
-        });
-
-      if (error) throw error;
-    }
-
-    res.json({ success: true });
-  } catch (error) {
-    console.error('Error saving file metadata:', error);
-    res.status(500).json({ error: "Error saving file metadata" });
-  }
-});
-
-// Authentication Endpoints
-router.post("/auth/login", async (req, res) => {
-  const { email, password, rememberMe } = req.body;
-
-  try {
-    const { data: user, error } = await supabase
-      .from('users')
-      .select('*')
-      .eq('email', email)
-      .single();
-
-    if (error || !user) {
-      return res.status(401).json({ error: "Invalid credentials" });
-    }
-
-    const isValidPassword = await bcrypt.compare(password, user.password_hash);
-    if (!isValidPassword) {
-      return res.status(401).json({ error: "Invalid credentials" });
-    }
-
-    const token = uuidv4();
-    const expiresAt = new Date(Date.now() + (rememberMe ? 14 : 1) * 24 * 60 * 60 * 1000);
-
-    await supabase.from('sessions').insert([{ user_id: user.id, token, expires_at: expiresAt }]);
-
-    res.json({ 
-      token, 
-      expiresIn: rememberMe ? '14d' : '24h',
-      user: { name: user.name, email: user.email }
-    });
-  } catch (error) {
-    console.error('Login error:', error);
-    res.status(500).json({ error: "Error during login" });
-  }
-});
-
-router.post("/auth/verify-email", async (req, res) => {
-  const { email, code } = req.body;
-
-  try {
-    const { data: user, error } = await supabase
-      .from('users')
-      .select('*')
-      .eq('email', email)
-      .single();
-
-    if (error || !user) {
-      return res.status(404).json({ error: "User not found" });
-    }
-
-    const { data: verificationRecord } = await supabase
-      .from('email_verification_codes')
-      .select('*')
-      .eq('user_id', user.id)
-      .eq('code', code)
-      .single();
-
-    if (!verificationRecord) {
-      return res.status(400).json({ error: "Invalid or expired verification code" });
-    }
-
-    await supabase.from('email_verification_codes').delete().eq('user_id', user.id);
-
-    res.json({ message: 'Email verified. You can now log in.' });
-  } catch (error) {
-    console.error('Email verification error:', error);
-    res.status(500).json({ error: "Error during email verification" });
-  }
-});
-
-// File Type Checking Endpoint
-router.get("/files/check-type", async (req, res) => {
-  const token = req.headers.authorization?.split(' ')[1];
-  if (!token) {
-    return res.status(401).json({ error: "No authentication token provided" });
-  }
-
-  const { fileId } = req.query;
-  if (!fileId) {
-    return res.status(400).json({ error: "File ID is required" });
-  }
-
-  // Parse fileId as integer to ensure it's a valid number
-  const parsedFileId = parseInt(fileId, 10);
-  if (isNaN(parsedFileId)) {
-    return res.status(400).json({ error: "Invalid file ID format" });
-  }
-
-  try {
-    // Check for localization data
-    const { data: localizationData, error: localizationError } = await supabase
-      .from('localization')
-      .select('id')
-      .eq('file_id', parsedFileId)
-      .limit(1);
-
-    if (localizationError) {
-      console.error('Error checking localization data:', localizationError);
-    }
-
-    // Check for designation data
-    const { data: designationData, error: designationError } = await supabase
-      .from('designation')
-      .select('*')
-      .eq('file_id', parsedFileId)
-      .single();
-
-    if (designationError && designationError.code !== 'PGRST116') {
-      console.error('Error checking designation data:', designationError);
-    }
-
-    // Check for test selection data
-    const { data: testSelectionData, error: testSelectionError } = await supabase
-      .from('test_selection')
-      .select('*')
-      .eq('file_id', parsedFileId)
-      .single();
-
-    if (testSelectionError && testSelectionError.code !== 'PGRST116') {
-      console.error('Error checking test selection data:', testSelectionError);
-    }
-
-    // Check for stimulation data
-    const { data: stimulationData, error: stimulationError } = await supabase
-      .from('stimulation')
-      .select('*')
-      .eq('file_id', parsedFileId)
-      .single();
-
-    if (stimulationError && stimulationError.code !== 'PGRST116') {
-      console.error('Error checking stimulation data:', stimulationError);
-    }
-
-    res.json({
-      hasLocalization: localizationData && localizationData.length > 0,
-      hasDesignation: !!designationData,
-      hasTestSelection: !!testSelectionData,
-      hasStimulation: !!stimulationData,
-      designationData: designationData ? {
-        designation_data: designationData.designation_data,
-        localization_data: designationData.localization_data
-      } : null,
-      testSelectionData: testSelectionData ? {
-        tests: testSelectionData.tests,
-        contacts: testSelectionData.contacts
-      } : null,
-      stimulationData: stimulationData ? {
-        stimulation_data: stimulationData.stimulation_data,
-        plan_order: stimulationData.plan_order,
-        is_mapping: stimulationData.is_mapping
-      } : null
-    });
-  } catch (error) {
-    console.error('Error in check-type endpoint:', error);
-    res.status(500).json({ error: "Error checking file type" });
-  }
-});
-
-// Localization Data Retrieval Endpoint
-router.get("/files/localization", async (req, res) => {
-  const token = req.headers.authorization?.split(' ')[1];
-  if (!token) {
-    return res.status(401).json({ error: "No authentication token provided" });
-  }
-
-  const { fileId } = req.query;
-  if (!fileId) {
-    return res.status(400).json({ error: "File ID is required" });
-  }
-
-  // Parse fileId as integer to ensure it's a valid number
-  const parsedFileId = parseInt(fileId, 10);
-  if (isNaN(parsedFileId)) {
-    return res.status(400).json({ error: "Invalid file ID format" });
-  }
-
-  try {
-    const { data: localizationData, error } = await supabase
-      .from('localization')
-      .select(`
-        id, contact, tissue_type, file_id,
-        electrode:electrode_id(id, label, description, contact_number, type),
-        region:region_id(id, name)
-      `)
-      .eq('file_id', parsedFileId);
-
-    if (error) {
-      console.error('Error fetching localization data:', error);
-      return res.status(500).json({ error: "Error fetching localization data" });
-    }
-
-    res.json(localizationData || []);
-  } catch (error) {
-    console.error('Error in localization endpoint:', error);
-    res.status(500).json({ error: "Error fetching localization data" });
-  }
-});
-
-// Signup Endpoint
-router.post("/auth/signup", async (req, res) => {
-  const { email, name, password } = req.body;
-
-  try {
-    // Check if user already exists
-    const { data: existingUser, error: checkError } = await supabase
-      .from('users')
-      .select('id')
-      .eq('email', email)
-      .single();
-
-    if (checkError && checkError.code !== 'PGRST116') {
-      console.error('Error checking existing user:', checkError);
-      return res.status(500).json({ error: "Error during signup" });
-    }
-
-    if (existingUser) {
-      return res.status(400).json({ error: "User with this email already exists" });
-    }
-
-    // Hash the password
-    const salt = await bcrypt.genSalt(10);
-    const hashedPassword = await bcrypt.hash(password, salt);
-
-    // Create new user
-    const { data: user, error: insertError } = await supabase
-      .from('users')
-      .insert([{ email, name, password_hash: hashedPassword }])
-      .select();
-
-    if (insertError) {
-      console.error('Error creating user:', insertError);
-      return res.status(500).json({ error: "Error creating user account" });
-    }
-
-    // Generate verification code
-    const verificationCode = Math.floor(100000 + Math.random() * 900000).toString();
-
-    // Store verification code
-    const { error: codeError } = await supabase
-      .from('email_verification_codes')
-      .insert([
-        { 
-          user_id: user[0].id, 
-          code: verificationCode, 
-          expires_at: new Date(Date.now() + 15 * 60 * 1000) 
-        }
-      ]);
-
-    if (codeError) {
-      console.error('Error storing verification code:', codeError);
-      return res.status(500).json({ error: "Error during signup process" });
-    }
-
-    res.json({ 
-      success: true, 
-      message: "User created successfully", 
-      verificationCode 
-    });
-  } catch (error) {
-    console.error('Signup error:', error);
-    res.status(500).json({ error: "Error during signup process" });
-  }
-});
-
-export default router;
